test(graph): Refactor mock printf into src/utils central utility and update Makefile (resolves #432)
