+++ src/graph_traversals/dijkstra.c
@@ -243,7 +243,7 @@ void add_edge_directed(weightedGraph* graph, int src, int dest, int wt)
     if (!graph)
         return;
 
-    if (src < 0 || src >= graph->V || dest < 0 || dest >= graph->V || wt < 0)
+    if (src < 0 || src >= graph->V || dest < 0 || dest >= graph->V)
     {
         printf("Invalid edge: %d -> %d with weight %d\n", src, dest, wt);
         return;
