+++ tests/graph_traversals/test_topological_sort.c
@@ -5,24 +5,7 @@
 #include <string.h>
 #include "graph_traversals.h"
 
-// Buffer to capture printf outputs from the algorithm
-static char g_printf_buf[4096];
-static int g_printf_len = 0;
-
-void reset_printf_buf()
-{
-    g_printf_buf[0] = '\0';
-    g_printf_len = 0;
-}
-
-int mock_printf(const char* format, ...)
-{
-    va_list args;
-    va_start(args, format);
-    g_printf_len += vsnprintf(g_printf_buf + g_printf_len, sizeof(g_printf_buf) - g_printf_len, format, args);
-    va_end(args);
-    return 0;
-}
+#include "mock_printf.h"
 
 // Redirect printf to our mock
 #define printf mock_printf
